Set clock with useState and refactor code
