--- src/components/Avatar/index.js
@@ -1,10 +1,0 @@
-export default function Avatar(props) {
-    return (
-        <div className="Avatar">
-            <img
-                src={props.author.avatarUrl}
-                alt={props.author.name}
-            />
-        </div>
-    );
-}
