+++ src/components/DucAnh/UserInfo/index.js
@@ -1,0 +1,12 @@
+import Avatar from "components/DucAnh/Avatar";
+
+export default function UserInfo(props) {
+  return (
+    <div className="UserInfo">
+      <Avatar author={props.author} />
+      <div className="UserInfo-name">
+        {props.author.name}
+      </div>
+    </div>
+  );
+}
