+++ src/components/DucAnh/Welcome/index.js
@@ -1,3 +1,3 @@
 export default function Welcome(props) {
-    return <h1> Hello, {props.name} </h1>
+  return <h1> Hello, {props.name} </h1>
 }
